Resolved codesmells
